contact-messenger-bot-functions/pyproject.toml
[tool.poetry]
name = "contact-messenger-bot-functions"
version = "0.1.0"
description = ""
authors = ["Robert Colfin <rcolfin@hotmail.com>"]
readme = "README.md"
packages = [
    { include = "contact_messenger_bot/" }
]

[tool.poetry.dependencies]
python = "^3.13"
functions-framework = "^3.8.2"
google-cloud-logging = "^3.11.3"
google-cloud-storage = "^2.18.2"
gunicorn = "^23.0.0"
contact-messenger-bot-api = {path = "../contact-messenger-bot-api"}

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"
ruff = "^0.6.9"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
pytest-ruff = "^0.4.1"
pytest-mypy = "^0.10.3"

[tool.pytest.ini_options]
addopts = "--doctest-modules --mypy --ruff --ruff-format"

[tool.mypy]
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = "google.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "google_auth_oauthlib.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "googleapiclient.*"
ignore_missing_imports = true

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "venv",
]

line-length = 120
indent-width = 4
target-version = "py39"

[tool.ruff.lint]
select = [
  "F",
  "E",
  "W",
  "C90",
  "I",
  "N",
  "UP",
  "YTT",
  "ANN",
  "ASYNC",
  "S",
  "BLE",
  "FBT",
  "B",
  "A",
  "C4",
  "DTZ",
  "T10",
  "EM",
  "EXE",
  "FA",
  "ICN",
  "G",
  'INP',
  'PIE',
  "T20",
  'PYI',
  'PT',
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SLOT",
  "SIM",
  "TID",
  "TCH",
  "INT",
  "PTH",
  "ERA",
  "PD",
  "PGH",
  "PL",
  "TRY",
  "FLY",
  "PERF",
  "RUF"
]
ignore = [
  "N999",
  "FBT002",
  "FBT001",
  "ANN101",
  "S101", # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
  "SIM102", # sometimes it's better to nest
  "UP038" # Checks for uses of isinstance/issubclass that take a tuple
          # of types for comparison.
          # Deactivated because it can make the code slow:
          # https://github.com/astral-sh/ruff/issues/7871
]
# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []
# The fixes in extend-unsafe-fixes will require
# provide the `--unsafe-fixes` flag when fixing.
extend-unsafe-fixes = [
    "UP038"
]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
force-single-line = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[build.env]]
    name = "GOOGLE_FUNCTION_TARGET"
    value =  "function"